requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
watchdog
ollama
python-dotenv